Adding --config-debug switch to save generated XML config file

+++ src/Command.php
@@ -225,9 +225,18 @@ class SolanoLabs_PHPUnit_Command
             SolanoLabs_PHPUnit_JsonReporter::writeJsonToFile($config->outputFile, $jsonData);
         }
 
+        // Save generated configuration XML file?
+        if ($config->configDebug) {
+            if (!empty(getenv('TDDIUM_TEST_EXEC_ID')) && !empty(getenv('TDDIUM_SESSION_ID'))) {
+                // Attach configuration XML to build page
+                rename($tempFile, getenv('HOME') . '/results/' . getenv('TDDIUM_SESSION_ID') . '/session/phpunit-' . getenv('TDDIUM_TEST_EXEC_ID') . '.xml');
+            } else {
+                echo("# XML phpunit configuration file: " . $tempFile . "\n");
+            }
+        } else {
+            unlink($tempFile);
+        }
 
-        // Delete temporary XML file
-        unlink($tempFile);
         return $exitCode;
     }
 
@@ -244,6 +253,7 @@ class SolanoLabs_PHPUnit_Command
         echo("   --tddium-output-file <file>  Can also be set with \$TDDIUM_OUTPUT_FILE environment variable\n");
         echo("   --ignore-exclude             Ignore <exclude/> child nodes of <testsuite/>.\n");
         echo("   --split                      Run tests one test file per process.\n");
+        echo("   --config-debug               XML configuration passed to phpunit will not be deleted.\n");
         echo("   -h|--help                    Prints this usage information.\n");
         echo(" * Any other supplied options will be passed on to phpunit\n");
     }
